refactor: split analysis bugs page modules
